move smoke to tools
