+++ tools/smoke/.gitignore
@@ -1,0 +1,1 @@
+grafana-agent-smoke
