+++ tools/smoke/internal/smoke.go
